+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/index.ts
@@ -52,7 +52,8 @@ export interface MyWidgetPreviewProps {
     style: string;
     styleObject?: CSSProperties;
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     valueAttribute: string;
     mywidgetValue: string;
     valueExpression: string;
@@ -125,7 +126,8 @@ export interface MyWidgetPreviewProps {
     style: string;
     styleObject?: CSSProperties;
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     valueAttribute: string;
     mywidgetValue: string;
     valueExpression: string;
@@ -189,7 +191,8 @@ export interface MyWidgetContainerProps {
 
 export interface MyWidgetPreviewProps {
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     valueAttribute: string;
     mywidgetValue: string;
     valueExpression: string;