+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/list-files.ts
@@ -22,7 +22,8 @@ export interface MyWidgetContainerProps {
 
 export interface MyWidgetPreviewProps {
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     actions: ActionsPreviewType[];
 }
 `;