+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/metadata-association.ts
@@ -25,7 +25,8 @@ export interface MyWidgetPreviewProps {
     style: string;
     styleObject?: CSSProperties;
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     data: {} | { caption: string } | { type: string } | null;
     metaReference: string;
     metaReferenceSet: string;