+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/selection.ts
@@ -30,7 +30,8 @@ export interface MyWidgetPreviewProps {
     style: string;
     styleObject?: CSSProperties;
     readOnly: boolean;
-    renderMode?: "design" | "xray" | "structure";
+    renderMode: "design" | "xray" | "structure";
+    translate: (text: string) => string;
     selectionAll: "None" | "Single" | "Multi";
     selectionSingleMulti: "Single" | "Multi";
     selectionMulti: "Multi";